app/requirements.txt
pymongo==4.8.0
neo4j==5.22.0
python-dotenv==1.0.1
pandas==2.2.3
matplotlib==3.9.2
seaborn==0.13.2
streamlit==1.40.0
